+++ common/src/main/java/com/slymask3/instantblocks/network/PacketHelper.java
@@ -2,6 +2,7 @@ package com.slymask3.instantblocks.network;
 
 import com.slymask3.instantblocks.block.InstantBlock;
 import com.slymask3.instantblocks.block.entity.*;
+import com.slymask3.instantblocks.builder.BuildSound;
 import com.slymask3.instantblocks.network.packet.InstantPacket;
 import com.slymask3.instantblocks.network.packet.client.*;
 import com.slymask3.instantblocks.network.packet.server.*;
@@ -36,7 +37,7 @@ public class PacketHelper {
     }
 
     public static void handleSound(SoundPacket message, Player player) {
-        for(Helper.BuildSound buildSound : message.buildSounds) {
+        for(BuildSound buildSound : message.buildSounds) {
             if(buildSound.getBreakSound() != null || buildSound.getPlaceSound() != null) {
                 ClientHelper.showParticles(player, buildSound.getBlockPos(), ClientHelper.Particles.PLACE_BLOCK);
             }